Improve error handling in bin/setup
In the EXIT signal trap, the command displays an error message if part of the set-up fails. However, to determine whether a command had failed, it was looking explicitly for an exit status of 1, which does not account for the various error-indicating exit statuses a command may return. This changes it to simply look for a non-zero status.